pal-389-moved null user service to services, updated config
