+++ user-service/src/main/java/uk/gov/gchq/palisade/service/user/service/NullUserService.java
@@ -13,12 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.gov.gchq.palisade.service.user.impl;
+package uk.gov.gchq.palisade.service.user.service;
 
 import uk.gov.gchq.palisade.User;
 import uk.gov.gchq.palisade.UserId;
 import uk.gov.gchq.palisade.service.user.exception.NoSuchUserIdException;
-import uk.gov.gchq.palisade.service.user.service.UserService;
 
 public class NullUserService implements UserService {
