+++ user-service/src/main/java/uk/gov/gchq/palisade/service/user/web/UserController.java
@@ -17,6 +17,7 @@ package uk.gov.gchq.palisade.service.user.web;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.core.io.Resource;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -43,7 +44,7 @@ public class UserController {
     private static final Logger LOGGER = LoggerFactory.getLogger(UserController.class);
     private UserService service;
 
-    public UserController(final UserService service) {
+    public UserController(@Qualifier("userService") final UserService service) {
         this.service = service;
     }
